+++ tests/ui_tests.rs
@@ -5,4 +5,5 @@ fn ui() {
     t.compile_fail("tests/ui/ambiguity.rs");
     t.compile_fail("tests/ui/recursion.rs");
     t.compile_fail("tests/ui/ws_cycle.rs");
+    t.compile_fail("tests/ui/bounds.rs");
 }